[nrf fromtree] Bsim: Bluetooth: Mesh: Add multicast bridge tests
Adds two new tests for the Subnet Bridge feature, testing group and
virtual addresses as destinations. The tests are based on the existing
`brg_simple.sh` test, with the addition of a new node and a new subnet
to make sure that messages are bridged to multiple subnets, but only
subnets that have a corresponding entry in the bridging table.

Signed-off-by: Håvard Reierstad <haavard.reierstad@nordicsemi.no>
(cherry picked from commit 0b35b38ef59b81040adc5014b9f82dd3863144ba)
Signed-off-by: Håvard Reierstad <haavard.reierstad@nordicsemi.no>